infrastructure/gcp/integrations/database/bigquery/bigquery/requirements.txt
google-cloud-storage==2.1.0
google-cloud-bigquery==2.32.0
pandas==1.4.0
pyodbc==4.0.32
